Add files via upload
I choose MLP classifier to do multi-label classification for otto group problem. Currently, the best I get is 0.79 accuracy, which it's not that good. I tried resampling to balance the data we have but it seems not very helpful. Besides, increasing complexity(more hidden layers) is also not good, so I just upload this file for now and if you guys find a way to improve the model, please feel free to modify it.